Added Jackson support for JSON parsing. Upreved for latest Markov changes

+++ src/main/java/com/bryanreinero/platypus/generator/RandIntervalGenerators.java
@@ -47,8 +47,8 @@ public class RandIntervalGenerators {
         switch (  spec.getType()  ) {
             case intType:
                 return new RandomInterval<Integer>() {
-                    private final int lowerBound = spec.getMin();
-                    private final int upperBound = spec.getMax( );
+                    private final int lowerBound = (Integer)spec.getMin();
+                    private final int upperBound = (Integer) spec.getMax();
 
                     public Integer getNext(){
                         int i;
